+++ registry/decoder.go
@@ -1,0 +1,62 @@
+package registry
+
+import (
+	"context"
+	"encoding/binary"
+	"fmt"
+
+	"github.com/hamba/avro/v2"
+)
+
+type Decoder struct {
+	client *Client
+	api    avro.API
+}
+
+// ExtractSchemaIDFromPayload extrapolates the schema id from a payload composed
+// of raw bytes containing a magic bytes, 4 bytes representing the schema encoding,
+// and the remaining payload being encoded with avro, as described in
+// https://docs.confluent.io/3.2.0/schema-registry/docs/serializer-formatter.html#wire-format .
+func extractSchemaIDFromPayload(payload []byte) (int, error) {
+	if len(payload) < 5 {
+		return 0, fmt.Errorf("payload too short to contain avro header")
+	}
+	if payload[0] != 0 {
+		return 0, fmt.Errorf("magic byte value is %d, different from 0", payload[0])
+	}
+	return int(binary.BigEndian.Uint32(payload[1:5])), nil
+}
+
+// NewDecoder shall return a new decoder given a client
+func NewDecoder(client *Client) *Decoder {
+	return &Decoder{
+		client: client,
+		api:    avro.DefaultConfig,
+	}
+}
+
+// DeserializePayload takes in input a payload to be deserialized, extrapolates
+// its schema id, gets the related schema and uses it to unmarshal the payload.
+// The payload shall be formatted according to:
+// https://docs.confluent.io/3.2.0/schema-registry/docs/serializer-formatter.html#wire-format .
+func (d *Decoder) DeserializePayload(
+	ctx context.Context,
+	payload []byte,
+	target interface{},
+) error {
+	if len(payload) < 6 {
+		return fmt.Errorf("payload not containing data")
+	}
+
+	id, err := extractSchemaIDFromPayload(payload)
+	if err != nil {
+		return fmt.Errorf("unable to extract schema id from payload, error: %w", err)
+	}
+
+	schema, err := d.client.GetSchema(ctx, id)
+	if err != nil {
+		return fmt.Errorf("unable to obtain schema, error: %w", err)
+	}
+
+	return d.api.Unmarshal(schema, payload[5:], target)
+}
